vision_server/requirements.txt
tensorflow
keras
numpy
opencv-contrib-python
python-chess